+++ src/Symfony/Component/Mailer/Tests/Transport/Smtp/DummyStream.php
@@ -55,7 +55,21 @@ class DummyStream extends AbstractStream
         $this->commands[] = $bytes;
 
         if (str_starts_with($bytes, 'EHLO')) {
-            $this->nextResponse = '250 localhost';
+            $this->nextResponse = '250 localhost'."\r\n".'250-AUTH PLAIN LOGIN CRAM-MD5 XOAUTH2';
+        } elseif (str_starts_with($bytes, 'AUTH LOGIN')) {
+            $this->nextResponse = '334 VXNlcm5hbWU6';
+        } elseif (str_starts_with($bytes, 'dGVzdHVzZXI=')) {
+            $this->nextResponse = '334 UGFzc3dvcmQ6';
+        } elseif (str_starts_with($bytes, 'cDRzc3cwcmQ=')) {
+            $this->nextResponse = '535 5.7.139 Authentication unsuccessful';
+        } elseif (str_starts_with($bytes, 'AUTH CRAM-MD5')) {
+            $this->nextResponse = '334 PDAxMjM0NTY3ODkuMDEyMzQ1NjdAc3ltZm9ueT4=';
+        } elseif (str_starts_with($bytes, 'dGVzdHVzZXIgNTdlYzg2ODM5OWZhZThjY2M5OWFhZGVjZjhiZTAwNmY=')) {
+            $this->nextResponse = '535 5.7.139 Authentication unsuccessful';
+        } elseif (str_starts_with($bytes, 'AUTH PLAIN') || str_starts_with($bytes, 'AUTH XOAUTH2')) {
+            $this->nextResponse = '535 5.7.139 Authentication unsuccessful';
+        } elseif (str_starts_with($bytes, 'RSET')) {
+            $this->nextResponse = '250 2.0.0 Resetting';
         } elseif (str_starts_with($bytes, 'DATA')) {
             $this->nextResponse = '354 Enter message, ending with "." on a line by itself';
         } elseif (str_starts_with($bytes, 'QUIT')) {