+++ src/Symfony/Component/Mailer/CHANGELOG.md
@@ -6,6 +6,8 @@ CHANGELOG
 
  * Add `MessageEvent::reject()` to allow rejecting an email before sending it
  * Change the default port for the `mailgun+smtp` transport from 465 to 587
+ * Add `$authenticators` parameter in `EsmtpTransport` constructor and `EsmtpTransport::setAuthenticators()`
+  to allow overriding of default eSMTP authenticators
 
 6.2.7
 -----